fix(mapmap): Remove unnecessary f32 cast in main.rs
Resolves clippy::unnecessary_cast warning in UIAction::SetSpeed handling.
Previously casting f32 to f32, which triggered CI failure.

+++ crates/mapmap/src/main.rs
@@ -2626,7 +2626,7 @@ impl App {
                 mapmap_ui::UIAction::Play => self.state.effect_animator.play(),
                 mapmap_ui::UIAction::Pause => self.state.effect_animator.pause(),
                 mapmap_ui::UIAction::Stop => self.state.effect_animator.stop(),
-                mapmap_ui::UIAction::SetSpeed(s) => self.state.effect_animator.set_speed(s as f32),
+                mapmap_ui::UIAction::SetSpeed(s) => self.state.effect_animator.set_speed(s),
                 _ => {
                     // Other actions might be handled elsewhere or are not yet implemented
                 }
